Land: Fix Y171 partial-magnitude speedF (f64->f32 HIO frame-rate constants)
The partial-magnitude (Y171) walk was speedF-RED at f17 (<=3 ULP) and pos_z-RED
at f27 (1 ULP). The standing hypothesis was a "jnt0 root-translate sub-ULP through
the WAITS<->WALK blend" (a calc_transform/Hermite frontier). That was wrong -- the
third such "FK frontier" that was really a plain f64-constant leak.

The daPy_HIO_move_c0 frame-rate fields (field_0x38=1.1, field_0x40=0.8, and the
atnMove/atnMoveB fields) are f32 members in the game, but the sim held them as
Python f64 literals fed straight into fp.fmuls/fadds/fdivs -- which do NOT quantize
their operands. setMoveAnime's rate f27r = f28 + f27*(f25*f3/f26 - f28) then rounded
1 ULP off, drifting the frame-ctrl phase -> the blend toe -> speedF. The full-
deflection walk never sees these (regime 3 collapses the rate to an exact value),
so only the partial regime was red. Fix: fp.f32(...) every such constant at
definition. walk_y171 -> 0 ULP offline (both tests) AND live (per-frame bit-exact).

Also quantize setSpeedAndAngleNormal's travel-step product to f32
(int(f32(param_1*dVar11)), matching decomp 2792/2796: an f32 multiply then s16
truncation) -- a latent decomp-faithfulness fix, inert on current seqs.

Re-diagnosed the two remaining reds via per-frame live comparison, overturning the
handoff's hypotheses: ebs is a backward-walk speedF residual (1-3 ULP scattered;
travel is bit-exact every frame), and waitturn is a walk-reentry speedF transient
(-192 ULP at the first MOVE frame after the pivot), not a turn-transient pos_x.
Both are the toe-FK/blend frontier, <=1 ULP pos_z, gameplay-neutral.

Offline land 182->184 pass; live gate walk_y171 0 ULP, no regression (only
ebs/waitturn red). KB: migrated the overturned Y171 claim to history/resolved-bugs;
refreshed land-sim/anim-engine/fp-faithfulness + test headers.

+++ tww_sim/land/land.py
@@ -299,10 +299,12 @@ class LandState:
                 else:
                     self.travel = cLib_addCalcAngleS(self.travel, self.target, self.F6, param_1, self.F4)
             else:
-                sVar6 = int(param_1 * dVar11)
+                # sVar6/sVar7 are s16 = (f32 product) truncated (decomp 2792/2796): quantize before
+                # int() -- an f64 product can truncate 1 unit off at a boundary, drifting the travel angle.
+                sVar6 = int(f32(param_1 * dVar11))
                 if sVar6 < 10:
                     sVar6 = 10
-                sVar7 = int(self.F4 * dVar11)
+                sVar7 = int(f32(self.F4 * dVar11))
                 if sVar7 < 1:
                     sVar7 = 1
                 self.travel = cLib_addCalcAngleS(self.travel, self.target, self.F6, sVar6, sVar7)
